requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
lgpio==0.2.2.0
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
python-engineio==4.9.1
python-socketio==5.11.2
rpi-lgpio==0.6
simple-websocket==1.0.0
Werkzeug==3.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4
smbus==1.1.post2